Added NCPF 11 with edge rule for non-turbines and int-based rules (untested)

+++ src/multiblock/configuration/underhaul/fissionsfr/PlacementRule.java
@@ -91,9 +91,9 @@ public class PlacementRule extends RuleContainer implements Searchable{
     }
     public Config save(Configuration parent, FissionSFRConfiguration configuration){
         Config config = Config.newConfig();
-        byte blockIndex = (byte)(configuration.blocks.indexOf(block)+1);
+        int blockIndex = configuration.blocks.indexOf(block)+1;
         if(parent!=null){
-            blockIndex = (byte)(parent.underhaul.fissionSFR.allBlocks.indexOf(block)+1);
+            blockIndex = parent.underhaul.fissionSFR.allBlocks.indexOf(block)+1;
         }
         switch(ruleType){
             case BETWEEN:
